feat(login): update login component
